Reorganize version management and use github to determine latest releases
