+++ support/vscode/koka.language-koka/package.json
@@ -3,7 +3,6 @@
   "displayName": "Koka Language",
   "description": "The official Koka programming language extension.",
   "version": "3.1.3",
-  "compilerVersion": "3.1.3",
   "publisher": "koka",
   "engines": {
     "vscode": ">=1.85.0"
@@ -211,6 +210,10 @@
         "command": "koka.selectCompiler",
         "title": "Koka: Select compiler path",
         "enablement": "koka.advancedCommands"
+      },
+      {
+        "command": "koka.downloadCompiler",
+        "title": "Koka: Download compiler"
       },
       {
         "command": "koka.selectTarget",
