+++ support/vscode/koka.language-koka/src/code-lens.ts
@@ -6,7 +6,7 @@ terms of the Apache License, Version 2.0. A copy of the License can be
 found in the LICENSE file at the root of this distribution.
 ---------------------------------------------------------------------------*/
 import * as vscode from "vscode"
-import { KokaConfig } from "./workspace";
+import { KokaConfig } from "./workspace-config";
 
 export class MainCodeLensProvider implements vscode.CodeLensProvider {
   private onDidChangeCodeLensesEmitter: vscode.EventEmitter<void> = new vscode.EventEmitter<void>()
