+++ support/vscode/koka.language-koka/src/extension.ts
@@ -9,7 +9,7 @@ import * as vscode from 'vscode'
 import * as path from 'path'
 import * as semver from "semver"
 
-import { KokaConfig } from './workspace'
+import { KokaConfig } from './workspace-config'
 import { CancellationToken, DebugConfiguration, DebugConfigurationProvider, ProviderResult, WorkspaceFolder } from 'vscode'
 import { KokaDebugSession } from './debugger'
 import { KokaLanguageServer } from './lang-server'
@@ -38,7 +38,7 @@ export async function activate(context: vscode.ExtensionContext) {
   if (semver.neq(prevVersion, kokaConfig.extensionVersion)) {
     // first time activation after an update/install
     await context.globalState.update('koka-extension-version', kokaConfig.extensionVersion)
-    await onUpdate(context, vsConfig, kokaConfig)
+    await onUpdate()
   }
 
   // only continue after here if the language server is enabled in the settings
@@ -47,10 +47,10 @@ export async function activate(context: vscode.ExtensionContext) {
     return
   }
 
-  // start the language service
+  // start the language server
   await startLanguageServer(context, vsConfig, kokaConfig, true /* allow install */)
-  if (!languageServer || !kokaConfig.hasValidCompiler()) {
-    console.log(`Koka: no valid compiler, don't start the language server (${kokaConfig.compilerPath})`)
+  if (!languageServer || !kokaConfig.versionManager.hasValidCompiler()) {
+    console.log(`Koka: no valid compiler, don't start the language server (${kokaConfig.versionManager.compilerPath})`)
     return;
   }
   console.log("Koka: language server started")
@@ -93,29 +93,29 @@ export async function activate(context: vscode.ExtensionContext) {
 // Check if the compiler has updated
 async function checkCompilerUpdate(context: vscode.ExtensionContext, vsConfig: vscode.WorkspaceConfiguration, kokaConfig: KokaConfig) {
   const prevCompilerVersion = await context.globalState.get('koka-compiler-version') as string ?? "1.0.0"
-  console.log(`Koka: check compiler update, previous: ${prevCompilerVersion}, current: ${kokaConfig.compilerVersion}`)
-  if (semver.neq(prevCompilerVersion, kokaConfig.compilerVersion)) {
+  console.log(`Koka: check compiler update, previous: ${prevCompilerVersion}, current: ${kokaConfig.versionManager.compilerVersion}`)
+  if (semver.neq(prevCompilerVersion, kokaConfig.versionManager.compilerVersion)) {
     // first time activation after an update/install of the compiler
-    await context.globalState.update('koka-compiler-version', kokaConfig.compilerVersion)
-    onCompilerUpdate(context, vsConfig, kokaConfig)
+    await context.globalState.update('koka-compiler-version', kokaConfig.versionManager.compilerVersion)
+    onCompilerUpdate()
   }
 }
 
 // Called after initial install and later updates of the compiler
-async function onCompilerUpdate(context: vscode.ExtensionContext, vsConfig: vscode.WorkspaceConfiguration, kokaConfig: KokaConfig) {
+async function onCompilerUpdate() {
   console.log("Koka: compiler is updated")
   await vscode.commands.executeCommand('koka.openSamples')
 }
 
 // Called after initial install and later updates of the extension
-async function onUpdate(context: vscode.ExtensionContext, vsConfig: vscode.WorkspaceConfiguration, kokaConfig: KokaConfig) {
+async function onUpdate() {
   console.log("Koka: extension is updated")
   await vscode.commands.executeCommand('koka.whatsnew')
   console.log("Koka: end of onUpdate")
 }
 
 // Clear all global state (for development)
-async function clearGlobalState(context: vscode.ExtensionContext, vsConfig: vscode.WorkspaceConfiguration, kokaConfig: KokaConfig) {
+async function clearGlobalState(context: vscode.ExtensionContext) {
   console.log("Koka: clear global state")
   await context.globalState.update("koka-compiler-version", undefined);               // last seen compiler version
   await context.globalState.update("koka-extension-version", undefined);              // last seen extension version
@@ -136,11 +136,11 @@ async function startLanguageServer(context: vscode.ExtensionContext,
     await stopLanguageServer(context)
   }
 
-  if (!kokaConfig.hasValidCompiler()) {
+  if (!kokaConfig.versionManager.hasValidCompiler()) {
     // update compiler paths and potentially install a fresh compiler
-    await kokaConfig.updateCompilerPaths(context, vsConfig, allowInstall)
-    if (!kokaConfig.hasValidCompiler()) {
-      console.log(`Koka: compiler is not functional: tried initializing from path(s): ${kokaConfig.compilerPaths.join(", ")}`)
+    await kokaConfig.versionManager.updateCompilerPaths(allowInstall)
+    if (!kokaConfig.versionManager.hasValidCompiler()) {
+      console.log(`Koka: compiler is not functional: tried initializing from path(s): ${kokaConfig.versionManager.compilerPaths.join(", ")}`)
       return false
     }
   }
@@ -176,7 +176,7 @@ function createBasicCommands(context: vscode.ExtensionContext, vsConfig: vscode.
     vscode.commands.registerCommand('koka.installCompiler', async () => {
       const noLanguageServer = (languageServer === null);
       await stopLanguageServer(context)
-      await kokaConfig.installCompiler(context, vsConfig)
+      await kokaConfig.versionManager.installCompiler()
       if (noLanguageServer) {
         // if this is the first time the compiler is installed, we need to reload
         return vscode.window.showErrorMessage('Reload VS Code to start the Koka language service with the new compiler')
@@ -190,14 +190,14 @@ function createBasicCommands(context: vscode.ExtensionContext, vsConfig: vscode.
     // Uninstall
     vscode.commands.registerCommand('koka.uninstallCompiler', async () => {
       await stopLanguageServer(context)
-      await kokaConfig.uninstallCompiler(context, vsConfig)
+      await kokaConfig.versionManager.uninstallCompiler()
       await vscode.commands.executeCommand("koka.restartLanguageServer")  // shows progress
       // await startLanguageServer(context,vsConfig,kokaConfig,false)
     }),
 
     // Clear global state
     vscode.commands.registerCommand('koka.clearState', async () => {
-      await clearGlobalState(context, vsConfig, kokaConfig)
+      await clearGlobalState(context)
     })
   )
 }
@@ -224,10 +224,10 @@ function createCommands(
   // select SDK
   context.subscriptions.push(
     vscode.commands.registerCommand('koka.selectCompiler', async () => {
-      kokaConfig.updateCompilerPaths(context, vsConfig, false);  // update with latest found paths
-      const path = await vscode.window.showQuickPick(kokaConfig.compilerPaths)
+      kokaConfig.versionManager.updateCompilerPaths(false);  // update with latest found paths
+      const path = await vscode.window.showQuickPick(kokaConfig.versionManager.compilerPaths)
       if (path) {
-        kokaConfig.setCompilerPath(path)
+        kokaConfig.versionManager.setCompilerPath(path)
       }
       if (selectSDKMenuItem) {
         selectSDKMenuItem.tooltip = `${path}`
@@ -246,9 +246,20 @@ function createCommands(
       }
     }),
 
+    vscode.commands.registerCommand('koka.downloadCompiler', async () => {
+      await kokaConfig.versionManager.getLatestKokaReleases();
+      const result = await vscode.window.showQuickPick(kokaConfig.versionManager.releases.map(r => r.version), {
+        placeHolder: "Select a Koka version to download"
+      });
+      if (result) {
+        kokaConfig.versionManager.setSelectedVersion(result);
+        await kokaConfig.versionManager.installCompiler();
+      }
+    }),
+
     // Open samples
     vscode.commands.registerCommand('koka.openSamples', () => {
-      kokaConfig.openSamples(context)
+      kokaConfig.openSamples()
     }),
 
     // Restart language server
@@ -329,7 +340,7 @@ function createCommands(
         options:
         {
           name: 'Koka interpreter',
-          shellPath: kokaConfig.compilerPath,
+          shellPath: kokaConfig.versionManager.compilerPath,
           cwd: kokaConfig.cwd,
           shellArgs: args
         }
